create util to handle version check
