+++ website/docs/docs/version-testing.md
@@ -3,14 +3,14 @@ title: "Version Testing"
 id: "version-testing"
 ---
 
-<VersionBlock firstVersion="1.8" lastVersion="1.10">
-  ## first version - 1.8 | last version - 1.10
+<VersionBlock firstVersion="1.6" lastVersion="1.9.1">
+  ## first version - 1.6 | last version - 1.9.1
 </VersionBlock>
 
-<!-- <VersionBlock firstVersion="2.2.1">
-  ## first version - 2.2.1
-</VersionBlock> -->
+<VersionBlock firstVersion="2.1">
+  ## first version - 2.1
+</VersionBlock>
 
-<!-- <VersionBlock firstVersion="1.10">
+ <VersionBlock firstVersion="1.10">
   ## first version - 1.10
-</VersionBlock> -->
+</VersionBlock>
